add django model inference
